+++ CHANGELOG.md
@@ -103,6 +103,7 @@
 * Use `==` for exact phrase search in Advanced Search for all full-text and term fields. Refs UIIN-2612.
 * Provide an instance `tenantId` to the PO line form when creating an order from the instance. Refs UIIN-2614.
 * Bump @folio/stripes-acq-components dependency version to 5.0.0. Refs UIIN-2620.
+* ECS: Check when sharing instance with source=MARC is complete before re-fetching it. Refs UIIN-2605.
 
 ## [9.4.12](https://github.com/folio-org/ui-inventory/tree/v9.4.12) (2023-09-21)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v9.4.11...v9.4.12)